+++ docs/release-notes.md
@@ -22,6 +22,7 @@ nav_order: 9
 - Support partitioning disk with mounted partitions
 - Support Proxmox VE
 - Support gzipped Akamai user_data
+- Support IONOS Cloud
 
 ### Changes